dumb switch stmts

+++ scripts/bootstrap/libraries.sh
@@ -4,22 +4,18 @@ dnc() {
 	FILE_INFO="$(file $1)"
 	FILE_LOCATION="$(dirname $1)"
 	FILE_NAME="$(basename $1)"
-	case "$FILE_INFO" in
-		'Zip')
-			mv "$FILE_LOCATION/$FILE_NAME{,.zip}" &&
-				unzip "$FILE_LOCATION/$FILE_NAME.zip" -od "$DIR_PROJECT/lib" &>/dev/null &&
-				rm -f "$FILE_LOCATION/$FILE_NAME.zip" &>/dev/null;
-		;;
 
-		'gzip')
-			mv "$FILE_LOCATION/$FILE_NAME{,.gz}" &&
-				gzip -d "$FILE_LOCATION/$FILE_NAME.gz" &>/dev/null
-		;;
-
-		*)
-			error "Could not decode unknown archive type at \"$(link_file $FILE_DEST)\"."
-		;;
-	esac
+	if [[ "$FILE_INFO" =~ 'Zip' ]]; then
+		mv "$FILE_LOCATION/$FILE_NAME"{,.zip} &&
+			unzip "$FILE_LOCATION/$FILE_NAME.zip" -od "$DIR_PROJECT/lib" &>/dev/null &&
+			rm -f "$FILE_LOCATION/$FILE_NAME.zip" &>/dev/null;
+	elif [[ "$FILE_INFO" =~ 'gzip' ]]; then
+		mv "$FILE_LOCATION/$FILE_NAME"{,.gz} &&
+			gzip -d "$FILE_LOCATION/$FILE_NAME.gz" &>/dev/null
+	else
+		error "Could not decode unknown archive type at \"$(link_file $FILE_DEST)\"."
+		return 1
+	fi
 
 	return 0;
 }
